HTTP GET / POST data added to recorded error.
The logger will now add GET or POST data to the dict that is sent to
errordite. This information appears in the exception report, but it is
mixed in with all the rest of the information - it's not shown in a
segregated block. That's an errordite issue. I may be possible to pre-
pend the dict key with 'GET_' or 'POST_', but that's for another commit.